Renamed Item object to task per ticket P1-2
